Ensure Unlimited Wanderers Patch wont remove too many instructions

+++ KaosesTweaks/Patches/BTUnlimitedWanderersPatch.cs
@@ -19,7 +19,10 @@ namespace KaosesTweaks.Patches
         private static IEnumerable<CodeInstruction> Transpiler(IEnumerable<CodeInstruction> instructions)
         {
             List<CodeInstruction> list = new List<CodeInstruction>(instructions);
-            list.RemoveRange(147, 3);
+            if (list.Count == 165)
+            {
+                list.RemoveRange(147, 3);
+            }
             return list.AsEnumerable<CodeInstruction>();
         }
